requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
idna==3.10
imbalanced-learn==0.13.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.6.3
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==10.4.0
propcache==0.3.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sklearn-compat==0.1.3
soundfile==0.13.1
threadpoolctl==3.6.0
ttkbootstrap==1.14.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wfdb==4.3.0
yarl==1.20.1
